9.4 Zadanie - FoodToDoor - korekty po uwagach mentora
