+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/foodtodoor/Customer.java
@@ -1,4 +1,4 @@
-package foodToDoor;
+package com.kodilla.good.patterns.foodtodoor;
 
 public class Customer {
     private String name;
