+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/foodtodoor/InformationService.java
@@ -1,4 +1,4 @@
-package foodToDoor;
+package com.kodilla.good.patterns.foodtodoor;
 
 public interface InformationService {
     void inform(Customer customer);
