+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/foodtodoor/OrderProcess.java
@@ -1,0 +1,6 @@
+package com.kodilla.good.patterns.foodtodoor;
+
+public interface OrderProcess {
+
+      OrderDto process(final OrderRequest orderRequest);
+}
