+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/foodtodoor/OrderService.java
@@ -1,4 +1,4 @@
-package foodToDoor;
+package com.kodilla.good.patterns.foodtodoor;
 
 public class OrderService {
     public boolean order (final Customer customer, final Supplier supplier, final int productQuantity){
