--- kodilla-good-patterns/src/main/java/foodToDoor/OrderProcess.java
@@ -1,10 +1,0 @@
-package foodToDoor;
-
-public interface OrderProcess {
-
-/*      void OrderServiceProcessor(final InformationService informationService,
-                                  final OrderService orderService,
-                                  final OrderRepository orderRepository);*/
-
-      OrderDto process(final OrderRequest orderRequest);
-}
